Abstract utilities into their own file.
No functional changees.
